dpkg-genchanges: Skip automatic debugging symbol packages
This is quite ugly as we are hardcoding a package name pattern here,
which as of now is distribution-specific. For the current automatic
debug package implementation, these contain an Auto-Built-Package
field, only present in the generated DEBIAN/control file, because
they do not exist in the debian/control file. We should eventually
detect that field and store it in the debian/files list file, and
then ignore the warning from dpkg-genchanges, but for now we'll just
quiesce the warning as it avoids unnecessarily confusing people
reading the output.