Dpkg::Vendor::Debian: Support getting the build path from the environment
Having dpkg-buildflags change its output depending on its current
working directory is not very friendly. We add a new environment
variable to be able to specify it so that we can control the output.

And use it from the test suite so make sure we always use the same
path regardless of where we execute the makefile snippets from.