dpkg: Use the conffile name instead of the real pathname it might refer to
When activating file triggers on conffile purge, use the conffile name
instead of the real pathname it might refer to. This fixes a segfault
when using --instdir, or when the conffile has been moved around and
replaced with a symlink to the target.

Regression introduced in 2f065fc151ac4635ae40bd769d1be810e1222e05.

Closes: #830267
Stable-Candidate: 1.17.x