Dpkg::Source::Archive: Add support for reproducible source packages
Clamp the mtime of the source files when building the tarball to the
first defined value from the source_date option, SOURCE_DATE_EPOCH
environment variable or time().

This makes sure the generated source tarballs always contain the same
mtime for files or directories that have been modified during the build.