dpkg-maintscript-helper: Support DPKG_ROOT for conffiles
As dpkg-maintscript-helper "works around known dpkg limitations in
maintainer scripts" it should behave as if it were a maintainerscript
and prepend DPKG_ROOT to all paths it uses.

This is preferred over having the conffile parameter(s) be prepended
with DPKG_ROOT by the maintainerscripts calling the helper as the helper
would then need to strip DPKG_ROOT from the path to verify if the
conffile belongs to the package – it also means that it magically works
for all callers instantly instead of requiring e.g. dh_installdeb to be
patched and packages to be rebuilt against a patched version.

Signed-off-by: Guillem Jover <guillem@debian.org>