build: Add support for switching man pages from roff to POD format
This makes editing, reading and translating the man page sources more
pleasant and obvious, as the markup is now way less obtrusive. For
translations the markup was some times leaking from po4a, as it did not
understand all roff macros, something that will stop happening now.

The man page sources required multiple fixes for issues that got
revealed during the conversion process, and massaging so that the
conversion would be easier to do automatically. These have been done in
previous commits.

The conversion also takes care of cleaning up markup artifacts due to
the way nroff works. To preserve the spacing in item entries, we had to
use S<>.

One issue from the conversion was preserving the usage of markup within
verbatim blocks. This is not officially supported by the POD format,
but there's a hack that can be used if our only target is man output.
The hack makes use of pod2man not reflowing the text, so that we trick
it into considering a paragraph as non-verbatim by making the first line
start on the first column with a zero-width space, and subsequent lines
start with a leading space, which will end up in the man page and
interpreted as verbatim blocks. It looks like this:

  ,---
  Some paragraph.
  Z<>
   Some verbatim C<block>,
   that supports B<markup>.
  `---

The notable differences are:

  - The generated files do not contain copyright nor license notices.
    This needs either support from pod2man(1) (say by adding support
    for a new =begin type of block), or perhaps get them injected by
    a post-processing script.
  - Many minus signs are output as hyphens, but this is pretty much
    a non-issue as most if not all roff processors have defaulted to
    make these equal for some time, so escaping them is not needed
    anymore, and is not really possible to do automatically in POD.

There are further cleanups that can be performed, but will be postponed
for a later point, so that they can be unfuzzied in block:

  - Use L<>, which got switched to be in bold instead of italic
    since Pod::Man 4.10 (2017-12-25) in perl 5.28.0.
  - Use F<> and C<>.
  - Switch from B<foo E<lt> bar> style mappings to B<< foo < bar >> form.