test: Delete MAKEFLAGS when testing make invocations
If the parent has been called with -jN then the MAKEFLAGS will contain
information to be used by the jobserver, but those file descriptors are
closed in the way, and make complains.

As we are testing the makefiles serially anyway, just remove the
MAKEFLAGS to get rid of the warnings.