dpkg-buildpackage: Enable -Jauto by default
The Debian build daemons have already been enabling parallel builds for
a some time now. This has the advantage that it should make the builds
faster, and that it gives similar conditions to local builders to the
ones on build daemon networks, helping to catch problems with parallel
build support in packages before the upload.

Closes: #842845
Suggested-by: Simon McVittie <smcv@debian.org>