dpkg: Fix use after free issue on error summary
We are releasing the dpkg database now after running the commands, which
means that the postponed error reporting summary was trying to print
messages that had already been freed from the database memory pool.

Duplicate the passed strings so that we are impervious to the database
life-cycle.

Regression introduced in commit 3404fd24ef8020b4d6dc17adb82d7e6c035d90dc.

Closes: #843874