dpkg-shlibdeps: Improve logic to identify packages owning a library
With things like merged-/usr, a system might have libraries that are
stored for example in /usr/lib but that dpkg knows under /lib. This
breaks some of the initial assumptions made in dpkg-shlibdeps.

We now scan all possible paths for a given library (instead of trying
to guess which one is the canonical one) and whenever we find a match
in the dpkg database, we also associate the package for the associated
realpath(). That way when a library is not properly identified, we can
fallback on looking if its realpath is known and be confident that if
the library was packaged, we did identify it correctly.

[guillem@debian.org:
 - Fold find_library_locations() into find_library().
 - Minor coding style fixes.
 - Squash the two separate commits and cleanup the commit message. ]

Closes: #843073
Signed-off-by: Guillem Jover <guillem@debian.org>