dpkg: Allow readlink(2) returning a size smaller than stat(2)
Some bogus filesystems do not return the actual symlink size in st_size,
which contradicts POSIX. But allowing the case where the returned size
is smaller than the one used to allocate memory is harmless, although
suspect. Let it through, but still print a warning so that users can
install stuff but are reminded they need to get a fixed filesystem in
place.

This has affected at least ecryptfs in the past and now file-based
encryption support in ext4 on Android N.

Reported-by: Jay Freeman <saurik@saurik.com>