dpkg-shlibdeps: Improve ELF ABI mismatch detector
The previous ELF ABI mismatch detector was very naïve, as the string
returned by «objdump -a» is a very simplistic representation of the
ELF ABI used.

Switch to our own ELF header parser, so that we can distinguish based
on the fields that define the object ABI.

This is still not enough, and we will have collisions with things such
as linux-i386 and hurd-i386, but most of the previously colliding
objects are now filtered.

This also makes us independent of objdump not supporting any unknown
ELF object ABI.

Closes: #849913