build: Disable disk pre-allocation by default
Contrary to what one would expect, this seems to be causing major issues
in several "modern" filesystems, as it collides with the heuristics and
optimizations that these try to perform.

Disable this by default, but let the builder enable it again in case
this is not a problem on certain systems.