Dpkg::Shlibs::Objdump: Fix read() error handling in get_format()
The rest of the code handles non-binary files (ELF in this case)
gracefully and ignores them, even though not very explicitly, as
objdump will emit a warning that might be difficult to decrypt.

We will still fail for other read failures that are not just
short-reads, as those imply some actual problem with the passed files.

Closes: #854536