Dpkg::Index: Add a new "unique_tuple_key" option to set_options()
When the option is true it will set a get_func_key to a function that
generates a unique tuple for that item. For CTRL_INDEX_SRC and CTRL_PKG_SRC
use the Package and Version fields, for CTRL_INDEX_PKG and CTRL_PKG_DEB use
the Package, Version and Architecture fields, all joined by "_" to index
the entries.

This behavior will become the default in dpkg 1.20.x. A deprecation
warning will be emitted now requesting to set the option to true or
to set a get_key_func options.

Prompted-by: Johannes Schauer <josch@debian.org>