dpkg: Fix conffile takeover handling during unpack on --root or diversions
The current code was trying to stat(2) a conffile without taking into
account a specified root directory, nor any possible diversions.

But the check based on the device and inode was pretty much unnecessary
because a simple string comparison is enough here, which as a bogus
solves both problems by reducing the code size.

Closes: #837051, #858004
Stable-Candidate: 1.17.x