dpkg-maintscript-helper: Print all unowned files from a directory
Add a new internal command that is not part of the public interface,
that takes care of checking whether a pathname is owned by a package.

This internal command has some logic to prevent external abuse, by
checking whether the parent has set a specific environment variable
with a token that changes per release.

The code will list all unowned files because xargs does not short-circuit
on the first command erroring out, as long these do not terminate with an
exit code of 255.

Closes: #813454, #860238
Based-on-patch-by: Bastien ROUCARIÈS <roucaries.bastien@gmail.com>