dpkg-shlibdeps: Preserve the order when scanning symbols/shlibs files
The code was getting all the possible shared library pathnames for the
wanted SONAME, but was not preserving the order carefully constructed
in find_library(), so we were overwriting symbols/shlibs information
when parsing multiple entries, and selecting the symbols/shlibs files
randomly based on the perl hash order.

This causes regressions when multiple packages provide the same
SONAME on different directories. An example would be libc6:amd64
and libc6-amd64:i386.

Fixes: commit a927295c93fb7a17742441aa863aaffcf4a351b5
Closes: #860979
Reported-by: Helmut Grohne <helmut@subdivi.de>