dpkg-genbuildinfo: Add duplicate prevention code for debian/files
We should do something similar to what dpkg-gencontrol is doing, by
preventing duplicated entries for the same file with different versions.

In this case, because the assumption is that there can ever only be one
buildinfo file for a «source» or «all» build, but possibly multiple for
arch-specific builds (from another build driver than dpkg-buildpackage),
we filter based on this.