s-s-d: Parse usernames and groupnames starting with digits correctly
We should not consider a username or groupname that starts with digits
as a valid uid or gid. When parsing integers we should parse the strings
fully and not consider any partial parsing to be correct.

Reported-by: Bodo Eggert <7eggert@online.de>
(cherry picked from commit 55c291c69bc9ee7c00731f4f5ffafd42673eb7d1)