Dpkg::Control::HashCore: Optimize trailing space matching on parse()
We remove the trailing space after the chomp, so that we cover the
common case of a single \n with chomp, and do not need to check for
trailing spaces more than once while parsing.

We preserve the chomp'ed string to be used for the Armor Header checks,
which have a different set of allowed whitespace, than what \s covers.