Dpkg::Source::Package: Call source format prerequisites method if present
We want to be able to check for prerequisites when loading the source
format module. That function used to be called import, which is a
misnomer here, and was not being invoked.

Rename it to the more appropriate prerequisites() and call it if it's
present from the module loading code.

Addresses: #877688