libcompat: Convert md5 module to use C99 int types
In commit 1f4d0354fc4a8c80c6463b9aaebd93822247341b we opted to map the
BSD int types to the C99 to avoid modifying the imported code. But we
had to modify it anyway, those types are legacy and we require the C99
ones for the rest of the codebase. So let's just switch to them, and
get rid of the type compatibility mapping.