Dpkg::Version: Fix bool overload behavior
The current bool overload has broken semantics, because it considers the
version "0" to be false.

The bool overload used to have sane semantics (equivalent to is_valid())
before commit 5b9f353b2940de751df47036608afbe71992d622, but there it got
changed to return the stringified version if it was valid, or undef
otherwise, to fix a problem within dpkg-shlibdeps, instead of properly
fixing the local-only problem in the tool. This makes the overload hard
to use, and broke existing callers from external projects.

We will emit a warning until dpkg 1.20.x to notify of the semantic change
in case there is code relying on the broken semantics. For fixed code the
warning can then be quiesced with:

  no warnings qw(Dpkg::Version::semantic_change::overload::bool);

Closes: #895004