dpkg-divert: Warn when using --add or --remove w/o --rename or --no-rename
Using --no-rename as the default optimizes for the wrong case, as that's
the exception, and while the safest option, it is needed only by packages
that are part of the pseudo-Essential set. It's also cumbersome for the
--local case.

We will emit a warning asking those to be explicit, so that we can switch
the default to --rename during the 1.20.x cycle.

Prompted-by: Paul Wise <pabs@debian.org>