Revert "Dpkg::Deps: Accept $archqual for add_provided_package() method"
This reverts commit 83272497c5be8c4e703ab179906cf904465fe775.

This commit introduced a regression in the author test suite. And there's
a patch by Johannes 'josch' Schauer <josch@mister-muffin.de> which should
be fixing this and other problems. If this is needed after all, we will
need to refactor the functions first to take a hash instead of a long list
of arguments.