Dpkg::Changelog::Parse: Stop using tail(1) to read the end of the file
Instead of relying on the tail command, simply read the end of the
file ourselves, assuming a packed set of 80 character lines, reading
4096 bytes before the end, implies around 51 lines, which is close
to the 40 lines currently used.

This should be both faster and should improve portability, because even
if we are using the POSIX -n option, some systems do not have a POSIX
compliant tail(1) on the default path, such as Solaris.

Analysis-by: James Clarke <jrtc27@debian.org>