dpkg-divert: Do not assume that diversions are always present
Several code paths assumed that the fsys nodes were only present when
they contained diversions, and no other nodes would be present, thus
did not make sure to check whether the divert struct members were valid.

With the switch to a single fsys underlying implementation now triggers
can vivify namenodes in the hash, so we need to check whether the members
are valid all the time.

Fixes: commit 2d7566140335d5338e5a98278d4df762936f67f4
Reported-by: Sven Joachim <svenjoac@gmx.de>
Bisected-by: Sven Joachim <svenjoac@gmx.de>