Dpkg::Version: Bump VERSION and improve semantic change description
Give context to the warning message, otherwise it's not clear what is
going on. If the users to not conditionalize the warning usage they will
get a perl warning about the unknown warning category itself, bump the
module version to be able to discern this change, and update documentation
to reflect this.