Dpkg::Source: Do not change patch permissions if not necessary
To be able to build a source tree, a user needs write permisions on it,
but not necessarily ownership of those files. We check the existing file
permissions and avoid changing them if not necessary, which helps in the
case where the user does not have ownership of those files, and they
were already present.

Closes: #898010
Naming-by: Julian Andres Klode <jak@debian.org>