dpkg-query: Run --list output through a pager if we are on a terminal
The current code, checks whether we are running on a terminal, and then
truncates the output fields so that it can fit within. This causes data
loss and can confuse the user depending where the cut point has happened.

Instead we remove that logic, and always redirect the outout through a
pager in case we are running on a terminal, which gives full information
and does not necessarily emit ugly output.

Closes: #898603