t-conffile-root-option: Override DPKG_INSTDIR and DPKG_ADMINDIR per target
We should not add a new --root option, as that means we will end up
unsing the admindir pre-configured in dpkg within that root directory,
which might not match what we expect.

Explicitly override both instdir and admindir, via target specific
variables.

This fixes dpkg builds with default configure pathnames.