Dpkg::Deps: Clarify arch-qualified dependency simplification
Dependency simplification can only really be done for metadata for which
we have all its context and information during the simplification
process. Anything that relies on the state of the dependencies cannot be
simplified. This means that any dependency that might change the
satisfiability due to the value of Multi-Arch field of the depended on
package cannot be simplified.

Clarify this in the function commends, and add new test cases to cover
this.

Prompted-by: Stuart Prescott <stuart@debian.org>