Dpkg::Deps::KnownFacts: Rename _evaluate_simple_dep to evaluate_simple_dep
This method is currently used by Dpkg::Deps::Simple, and is marked as
private by way of its underscore prefix. This will be a problem once
we split the modules into different files, as then we are not supposed
to be using private methods from external modules.

Rename the method, and document it explicitly as being private, while
only giving a partial prototype to deter its use.