libdpkg: Change pkg_name() and pkgbin_name() back to get non-const structs
This partially reverts commit cd96cca4b2cf83f46d6e289418ed06e3c2ef7066.

It is not safe to assume that the pkgname_arcqual member has been
populated before. The caller might have used a local variable filled
manually. So we should always check whether it is NULL and fill it
ourselves, before returning it.

In addition now that we have the const function variants we can turn
these back into non-const so that we can do the archqualified package
name generation, and cache it for later use.