libdpkg: Add new pkg_name() and pkgbin_name() const variants
Because C does not have a C++ equivalent for the mutable keyword, we
cannot mark the pkgname_archqual member as such. Instead we will add a
new couple of const functions that will fallback to return a non-freeing
string to be used mostly in error handling code paths in case there is
no cached member, as otherwise we'd be "leaking" those strings (just
generating new instances that will be released as part of the pool)
every time we call these functions.