test: Delete fixup lines from i18nspector output instead of emptying them
The previous regex was just turning the line into an empty one, but
we count the length to check whether there was any output. In this
particular case this did not trigger because we have a chomp, that
would get rid of the last empty line, which happened to be fine as
we currently only emit one such empty line per file. But this is not
future-proof, so let's fix it before it bites us.