s-s-d: Check whether standalone --pidfile use is secure
If we are only matching on the pidfile, which is owned by a non-root
user, and we are running as a root user then this is a security risk,
and the contents cannot be trusted, because the daemon might have been
compromised which would allow modifying the pid within.

If we are then calling start-stop-daemon as a privileged user, that
would allow acting on any PID in the system.

Prompted-by: Michael Orlitzky <michael@orlitzky.com>
Ref: https://redmine.kannel.org/issues/771