build: Perform Dpkg module substitution during CPAN module building
We had an empty do_perl_subst variable in the dist-cpan autotools
target that resulted in copying an empty file, but that was shadowed by
a subsequent copy of the original. So we were ending up with a pristine
non-substituted file.

We actually need to substitute the install paths during the CPAN module
building to preserve system and user settings at that point. Otherwise
the dpkg autotools configure settings might be completely out of touch
with the settings from the system where the CPAN module gets built.