libdpkg: Call the pager with «$SHELL -c» to respect POSIX
The PAGER environment variable is specified in POSIX as taking a pager
command and any number of arguments to be passed to «sh -c», so we have
to preserve that behavior or this breaks badly on systems where PAGER is
set that way.

This actually gets in the way of improving the pager command presence
detection, but we'll fix that later on, by making the scanning of the
environment variable smarter or by trying to execute the fallbacks.

Closes: #910009