libdpkg: Ignore SIGPIPE when setting up a pager
If the pager quits early, the parent will receive a SIGPIPE as the write
end of the pipe will not be available anymore. Instead we ignore SIGPIPE
and also EPIPE errors when writing to stdout.

Otherwise if we quit the pager early, the program will exit with an
error code.

Reported-by: Holger Levsen <holger@layer-acht.org>
Ref: #909754