Dpkg::Source::Package: Improve debian/source/format parsing and validation
Make the regex more strict and decompose it right away instead of doing
a second pass over it. Only initialize minor when we need to update the
format. And fix the error for an invalid format to stop referencing the
Format field, which might not be involved during the parsing.