libdpkg: Increase the pkg-hash bins size
We use a number that is close to the amount of packages currently present
in a Debian suite, so that installed and available packages do not add
tons of collisions.

The memory usage is «BINS * sizeof(void *)», so it will be 256 KiB on
32-bit systems and 512 KiB on 64-bit systems.

Update the code comment to reflect the above.