s-s-d: Add new fatalv() and fatale() and use the latter for system errors
We should not try to use errno, when the error has not been produced due
to a system error, as then errno might contain bogus or unrelated
information. Be explicit in the error handling by calling fatale() for
errno related errors, and fatal() for the ones that are disconnected
from it.