debian: Remove Origin and Bugs fields from control file
These fields have been there mostly to showcase their usage, which
should instead be dynamically added at build-time, and as a reminder
that this should be done.

But there is already a bug filed as a reminder (see #511160), they
are also now documented in the deb-src-control(5) man page, they are
actually wrong in many cases when present on a downstream, and they
generate nagging lintian warnings to boot.