libdpkg, dpkg: Print the dpkg database directory on access errors
There were several error messages involving the dpkg database, that were
ambiguous by referring to it as "dpkg status area", instead of printing
the affected pathname. The historical reason for not being more explicit
has been to keep the user a bit under the dark, because this has been an
internal database storage, and only those in the known would go there to
edit it by hand and similar.

This is a bit pointless though, because the Internet is full of
references to those directories, and other error messages just print
them anyway.

In addition, for users that want to diagnose what is the problem this is
not helpful, and the directory might have been changed at build or run
time with one of the options.

Closes: #883700