debian: Add Breaks on lsb-base due to s-s-d breaking killproc
There was a bug in the killproc function implementation that made it not
pass the --name argument, so it would end up with just the --pidfile.
The new security requirements from s-s-d made that in some cases, when
the usage was not secure, it would bail out with an error.

Closes: #923861