scripts: Switch Getopt::Long from bundling to bundling_values
This got introduced after a request so that we could switch our tools
to a safe subset of the bundling logic, so let's enable it now that it
is available in the minimum supported perl version.

Bundling is in general unsafe, and not future-proof, as new options
might gain arguments which would then get misinterpreted, we should
not really be using it. But bundling the value into the option is a
common pattern and should be allowed.