libdpkg: Add new C locale switch over support
This will make it possible to select either the current locale or a
specified one, currently only C is selectable.

We use this because the uselocale(3) POSIX API is recent, and the system
might not have it available, so we might need to fallback to use
setlocale(3) instead, which even though is not thread-safe, it's better
than nothing.