debian: Replace custom rule for 'configure' with call to dh_autoreconf
Having this custom rule to create the «configure» file means that it
might end up not being rebuilt if the timestamps of some other files
are off compared to the debian/changelog date. Using dh_autoreconf
guarantees that even if these timestamps got updated incorrectly,
the file will always be rebuilt as expected.

The historic rationale for not doing this before, even though this has
been considered a best packaging practice, was that because dpkg needs
explicit support for new architectures, which is the main reason to
get autotools updated to pull in new config.* files, we would get
reports so that we could add such support and release a new upstream
version. But this has never happened, and it's causing other problems.

Closes: #939516
Signed-off-by: Guillem Jover <guillem@debian.org>