dpkg: Use DPKG_ADMINDIR to set the admindir
This was done in the name of security, but it does not make much sense
as root should make sure to be running in a clean environment. This is
problematic as it makes the interface hard to understand, and produces
unexpected behavior as dpkg can be called from within itself, or as a
wrapper for dpkg-query, which will then not inherit the admindir value.

Closes: #900071