libcompat: Disarm libselinux setexecfilecon() declaration for libcompat-test
When building the selinux module as part of the libcompat-test library,
that takes care of making sure the libcompat code can compile all the
time, we are redeclaring the symbol which causes a compiler warning.

Remap the symbol while we include the libselinux headers and then remove
the remapping so that ours is used.

Warned-by: gcc -Wredundant-decls