dpkg-fsys-usrunmess: New program
This program undoes the merged-/usr-via-aliased-dirs mess done on
filesystems that have been installed anew with recent installers with
unfortunate defaults or systems migrated with the convert-usrmerge tool
from the usrmerge package.

Note: while dpkg.deb intends to contain no perl code, to be able to
handle upgrades properly, and to reduce the perl usage on minimal
systems, this one is made an exception as it is a script that does
not depend on anything non-core including libdpkg-perl (except for
libfile-fcntllock-perl checked at run-time), and is to be executed as
a one-off thing. And it is more important to give people that have
such broken systems recover from that, than keep the aforementioned
properties very strictly for this particular case.