build: Switch to preserving the whole environment within sudo
Trying to whitelist what might be needed is too cumbersome. The latest
instance was this breaking being run from within cowbuilder. Instead we
preserve the environment and do not have to worry about this anymore.

We still need to preserve LD_PRELOAD and LD_LIBRARY_PATH because these
get reset when going over a setuid-root barrier, such as sudo.

Prompted-by: Niels Thykier <niels@thykier.net>