build: Handle po4a --porefs difference between 0.58 and earlier versions
Before po4a 0.58 --porefs supported specifying a wrapping mode by
appending «,wrap» or «,nowrap», with the default being “nowrap”.
Starting with po4a 0.58 the default has changed to “wrap”, and these
modifiers are not accepted anymore, which causes a problem with
backwards compatibility as we cannot get consistent behavior across
different versions.

Detect this in configure and pass the «,wrap» modifier when needed
and supported, which should make it possible to build the man pages
again consistently in Debian buster and bullseye.